Remove server target, fix quarantine safety, open-source the project
Consolidates PDFSafe into a single-target Windows desktop application
and prepares it for public release.

Removed the optional FastAPI/Celery/PostgreSQL server (~22 files). Its
duplicated orchestration had silently drifted from the desktop engine
twice, and the second drift caused a bug no test caught because the
server suite was skipped in the desktop environment.

Quarantine no longer deletes the user's file. A malicious verdict now
renames it to <name>.quarantine, stripping the .pdf association without
destroying data — the false-positive rate is unmeasured, so an
irreversible action was the wrong default. Adds a restore path so
'mark as safe' undoes both halves.

Also fixed: re-quarantining identical content failed silently against a
read-only vault entry; a quarantine failure discarded a completed
analysis; history pruning leaked stored blobs; the daily AI token budget
was backed by Redis and therefore never applied on a desktop install.

Licensed under Apache-2.0 with third-party attribution in NOTICE,
including the LGPL position on Qt/PySide6. Adds SECURITY.md with a
threat model, CONTRIBUTING.md, and issue templates for false positives
and missed detections.

142 tests passing, mypy clean.

Detection thresholds remain unvalidated against a real-world corpus.

+++ src/pdfsafe/ai/triage.py
@@ -176,7 +176,7 @@ def _fuse(
     assert call.verdict is not None  # guarded by the caller
     ai = call.verdict
 
-    blended = int(round(AI_SCORE_WEIGHT * ai.risk_score + (1 - AI_SCORE_WEIGHT) * outcome.score))
+    blended = round(AI_SCORE_WEIGHT * ai.risk_score + (1 - AI_SCORE_WEIGHT) * outcome.score)
 
     verdict = ai.verdict
     has_critical = any(i.severity is Severity.CRITICAL for i in outcome.indicators)
